fix(app): Resolve image upload crash and frontend reference error
This commit addresses two issues related to the image upload functionality on the ingredients page:

1.  **Backend Crash on Session Update:** The application was crashing due to resource exhaustion when the large base64 data URL of an uploaded image was being persisted to the backend session store. This is fixed by stripping the `finalRecipeImageUrl` and `baseRecipeImageUrl` properties from the session update payload in `product-generator.tsx` before it is sent to the server.

2.  **Frontend ReferenceError:** A `ReferenceError` was occurring in the production build, likely due to a module initialization order issue. This is resolved by refactoring the `IngredientsStep` component. The `SubIngredientModal` and its related interfaces have been extracted from `ingredients-step.tsx` into their own separate file, `sub-ingredient-modal.tsx`, to ensure proper module isolation.

These changes together ensure that the image upload feature is robust, does not crash the server, and renders correctly on the frontend without errors.
